guard write
Guard write call, avoid ignored-return-value warning.

(This is not a false positive: a very subtle attack would consist of
filling up the filesystem so much that only a partial PID is written,
causing the wrong PID to be signaled later.)

(Note that, technically speaking, if only some of the buffer is
written we should retry the rest in a loop.  But in this case, that
seems exceedingly unlikely.)

Gbp-Pq: Name 0006-guard-write.patch